+++ src/main/resources/lang/en.yml
@@ -10,6 +10,7 @@ General:
   NotFound: "&e%backpack% &cdoes not exist."
   OverLimit: "&7Removed &c%removed% backpacks &7because you are only allowed to carry &e%limit% &7at once."
 BackPack:
+  ItemNotAllowed: "%backpack% &cdoes not let you store &e%item% &cinside."
   OnCraft:
     - "&8[&a+&8] &aYou have crafted a %backpack%"
   OnGive: "&7You have given &e%player% &7- &a%amt%x %backpack%"
